+++ src/qhexedit.h
@@ -346,8 +346,8 @@ private:
     void resetSelection(qint64 pos);            // set selectionStart and selectionEnd to pos
     void resetSelection();                      // set selectionEnd to selectionStart
     void setSelection(qint64 pos);              // set min (if below init) or max (if greater init)
-    int getSelectionBegin();
-    int getSelectionEnd();
+    qint64 getSelectionBegin();
+    qint64 getSelectionEnd();
 
     // Private utility functions
     void init();
